Entirely replace build source with the created release
Elixir releases contain everything needed to run the application inside the
release. As the app grows, duplicating all the source code, assets, etc.,
creates a very large slug, and it's easy to exceed Heroku's 500MB hard limit.

Instead, swap the entire build source code out with the release, so the release
sits in /app. This is a breaking change, as all configuration that points to the
path of the release needs updating, like the `Procfile` or `app.json`.